6.2 done
Extremely js solution.
The observation was that, even considering multiple disconnected trees in the input,
there are paths from a common centre of mass to each of the points we want to measure.
Get both those paths then sum the steps to each from their closest common ancestor.
The execution of that idea gets no points for data structure elegance.